server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
websockets==12.0
groq==0.4.1
python-multipart==0.0.6
numpy==1.26.3
httpx>=0.25.0
ffmpeg-python>=0.2.0
requests>=2.28.0